requirements.txt
feedparser>=6.0.0
requests>=2.31.0
python-dateutil>=2.8.0
pyyaml>=6.0.0
beautifulsoup4>=4.12.0
nltk>=3.8.0
textblob>=0.17.0